Moving misplaced commands out of AutoScoringCommandFactory
