+++ src/main/java/frc/robot/subsystems/drive/commands/AutoScoringCommandFactory.java
@@ -2,8 +2,6 @@ package frc.robot.subsystems.drive.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.constants.RobotConstants.CoralOuttakeConstants;
-import frc.robot.constants.RobotConstants.LiftConstants.CoralPreset;
 import frc.robot.subsystems.algae.AlgaeSubsystem;
 import frc.robot.subsystems.algae.commands.AlgaeCommandFactory;
 import frc.robot.subsystems.coralouttake.CoralOuttakeSubsystem;
@@ -11,8 +9,6 @@ import frc.robot.subsystems.coralouttake.commands.CoralOuttakeCommandFactory;
 import frc.robot.subsystems.gantry.commands.GantryCommandFactory;
 import frc.robot.subsystems.lift.LiftSubsystem;
 import frc.robot.subsystems.lift.commands.LiftCommandFactory;
-import java.util.function.BooleanSupplier;
-import java.util.function.Supplier;
 
 public class AutoScoringCommandFactory {
   private GantryCommandFactory gantryCommandFactory;
@@ -40,11 +36,6 @@ public class AutoScoringCommandFactory {
     this.algaeSubsystem = algaeSubsystem;
   }
 
-  public Command gantryAlignCommand(Supplier<CoralPreset> getPreset, BooleanSupplier getDsSide) {
-    return gantryCommandFactory.gantryPIDCommand(
-        () -> getPreset.get().getGantry(getDsSide.getAsBoolean()));
-  }
-
   public Command autoPlace() {
     return gantryCommandFactory
         .gantryDriftCommand()
@@ -65,28 +56,5 @@ public class AutoScoringCommandFactory {
         .andThen(
             new WaitCommand(0.1)
                 .deadlineFor(coralOuttakeCommandFactory.setIntakeVoltage(() -> 4).repeatedly()));
-  }
-
-  public Command placeTrough() {
-    return coralOuttakeCommandFactory
-        .setIntakeVoltage(() -> 12)
-        .repeatedly()
-        .until(() -> !coralOuttakeSubsystem.hasCoral())
-        .andThen(
-            new WaitCommand(0.1)
-                .deadlineFor(coralOuttakeCommandFactory.setIntakeVoltage(() -> 12).repeatedly()));
-  }
-
-  public Command algaeAutoPickup() {
-    return algaeCommandFactory
-        .setSolenoidState(() -> true)
-        .andThen(algaeCommandFactory.setMotorVoltages(() -> 4, () -> 4))
-        .repeatedly()
-        .until(() -> algaeSubsystem.hasAlgae());
-  }
-
-  public Command outtakeCoralCommand() {
-    return coralOuttakeCommandFactory.setIntakeVoltage(
-        () -> CoralOuttakeConstants.passiveSpeed * 12);
   }
 }
